move InternalContractFunctionSet to testutils
