+++ soroban-sdk/src/env.rs
@@ -256,32 +256,6 @@ enum EnvTestState {
     Contract,
 }
 
-/// Adapts a [`ContractFunctionSet`] into the function set the host dispatches
-/// native contract calls to.
-///
-/// Shared by contract registration and Wasm upload, both of which hand the host
-/// a native contract to dispatch to.
-#[cfg(any(test, feature = "testutils"))]
-pub(crate) struct InternalContractFunctionSet<T: ContractFunctionSet>(pub(crate) T);
-
-#[cfg(any(test, feature = "testutils"))]
-impl<T: ContractFunctionSet> internal::ContractFunctionSet for InternalContractFunctionSet<T> {
-    fn call(&self, func: &Symbol, env_impl: &internal::EnvImpl, args: &[Val]) -> Option<Val> {
-        let env = Env {
-            env_impl: env_impl.clone(),
-            test_state: EnvTestState::Contract,
-        };
-        self.0.call(
-            crate::Symbol::try_from_val(&env, func)
-                .unwrap_infallible()
-                .to_string()
-                .as_str(),
-            env,
-            args,
-        )
-    }
-}
-
 #[cfg(any(test, feature = "testutils"))]
 impl EnvTestState {
     fn config_mut(&mut self) -> &mut EnvTestConfig {
@@ -636,8 +610,8 @@ use crate::{
     testutils::{
         budget::Budget, cost_estimate::NetworkInvocationResourceLimits, default_ledger_info,
         Address as _, AuthSnapshot, AuthorizedInvocation, ContractFunctionSet, EventsSnapshot,
-        Generators, Ledger as _, MockAuth, MockAuthContract, Register, Snapshot,
-        SnapshotSourceInput, StellarAssetContract, StellarAssetIssuer,
+        Generators, InternalContractFunctionSet, Ledger as _, MockAuth, MockAuthContract, Register,
+        Snapshot, SnapshotSourceInput, StellarAssetContract, StellarAssetIssuer,
     },
     Bytes, BytesN, ConstructorArgs,
 };
@@ -655,6 +629,15 @@ use xdr::{LedgerEntry, LedgerKey, LedgerKeyContractData, SorobanAuthorizationEnt
 #[cfg(any(test, feature = "testutils"))]
 #[cfg_attr(feature = "docs", doc(cfg(feature = "testutils")))]
 impl Env {
+    /// Create an [Env] for a call into a native contract's function, where the
+    /// test state is unavailable because the code is running as the contract.
+    pub(crate) fn for_contract_fn(env_impl: internal::EnvImpl) -> Env {
+        Env {
+            env_impl,
+            test_state: EnvTestState::Contract,
+        }
+    }
+
     #[doc(hidden)]
     pub fn in_contract(&self) -> bool {
         self.env_impl.has_frame().unwrap()
